Create Adjusted Screens Script (DRAFT)
This contains the entire screen script, even elements that we do not want to change. The relevant section for our purpose starts at line 97 and goes until 372. This is just a draft version to test certain parameters and can be used to explore how to adjust screens, but it should not be used as a final version.