Add files via upload
These are three scripts that should be used to help create various screens for Stream Evil. 

The "Script" script contains a series of interactions that test the screens' functionality, making sure that dialogue boxes are moving to appropriate locations at the desired time, that stream chats can be added to the chat screen when desired, and that transitions work within the layers of Ren'Py.

You can copy/paste the Screens script as a replacement for the standard screens script in any Renpy game file. It will enable the use of a screen for the chat function, the streamer's details at the bottom of the screen, and some basic code to potentially use for the Discord scenes. It also contains a handful of window "styles" that will be used in the script.

The "Lists" script contains the lists of comments that can be added to the stream chat as well as the python functions and definitions that enable this functionality. It should also contain labels that are called when textbuttons are clicked from the stream chat screen.