Update script.rpy
I updated the first 50 lines with the character definitions and variables we'll be using. Then I changed the body of the script (lines 65ish through 183) to represent a basic version of the beginning of vignette 2. This script will be able to be tested when used alongside the "lists" and "screens" script here on GitHub as well as the images from the Assets folder in Google Drive.

I kept the previous script used by Justin to run tests because I think it could be useful for debugging purposes. I just shifted it to the bottom of the script and put it inside a label called "justinTests." If we ever want to use it to debug or test, we can just put the line "jump justinTests" near the start of the script to load that sequence.