Update Lists.rpy
I simplified the lists script. There were some functions (like addStoryBeat and removeChats) that could be deleted. I also added comments tha can appear at the beginning of the game and in the first two choices of vignette 2. The labels that they direct to can be found under the new script "targets."