Update screens.rpy
Adjusted code for the streamChat screen (lines 87-124). Made it so that new comments get added to the bottom of the screen instead of the top and added functionality so that the screen will scroll down to the bottom whenever a new chat is added. Weirdly, this code is not functioning 100% of the time. Still working on some bugs